### 2022-01-21 #4082 @whedon
[PRE REVIEW]: ALUES: An R package for Agricultural Land Use Evaluation System
**Submitting author:** @alstat (<a href="http://orcid.org/0000-0003-3784-8593">Al-Ahmadgaid Asaad</a>)
**Repository:** https://github.com/alstat/ALUES
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.1
