### 2022-02-21 #4082 @fabian-s
:wave: @nuest, :wave: @benmarwick,  :wave: @dlebauer

would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
