### 2022-03-07 #4082 @fabian-s
:wave: @jhollist @corybrunson @cole-brokamp

would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
