### 2022-03-14 alstat/ALUES#15 @corybrunson
clarify contributions of authors
Three authors are credited on the package and on the manuscript, but only one directly contributed code to the repo. Could each author's contributions be briefly (1 phrase) explained in a statement, either in the package (e.g. README or ALUES.R) or on the manuscript? This would address a specific review criterion. Thank you!

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/4228)._
