### 2022-03-14 alstat/ALUES#16 @corybrunson
what is "efficient enough"?
This might just be a typo, but i want to be sure. In the JOSS manuscript, you say in the "Speed" section, "ALUES is efficient enough in evaluating large sets of land units." Is there a domain-specific threshold these systems are expected to meet, or is this just a general statement that ALUES is fast?

If it's a generic statement, it would be helpful to know whether ALUES is competitive, i.e. how fast other systems are at the same task. If this information is not available, then perhaps the benchmark tests can be presented without a normative claim about their efficiency?

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/4228)._
