### 2022-03-14 alstat/ALUES#17 @corybrunson
statement of need in software documentation
The statement of need in the JOSS manuscript is very clear. However, one review criteria is that the software documentation itself also contain a statement of need. I don't expect it to be as long as that in the manuscript; a sentence or two in the README or manual could explain what need this package fills.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/4228)._
