### 2022-03-15 #4228 @corybrunson
@fabian-s FYI i checked roughly half the boxes on my checklist yesterday and today found them all unchecked. Moreover, i find myself now unable to check boxes. Do you have an idea of what might be wrong? I will try logging out and back in, etc.
