### 2022-03-15 alstat/ALUES#18 @corybrunson
clarify mismatch error behavior
I appreciate the thorough collection of vignettes written for this package, and in particular i'm glad that the documentation for `suit()` and [this vignette](https://alstat.github.io/ALUES/articles/a04_understanding_overall_suit.html) illustrate error behavior.

However, neither prints the error itself, and the documentation for `suit()` doesn't explain it. This kind of error arose quickly when i started experimenting with the functionality, so i would urge that it be detailed in both places—with an explanation in the documentation and e.g. with a printout of `banana_suit[["terrain"]]` in the vignette.

(I don't have an agricultural background, so please just let me know if this really should not be necessary for the target user base!)

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/4228)._
