### 2022-03-15 alstat/ALUES#19 @corybrunson
unnecessary dependencies?
I'm not clear on whether the ggmap and raster packages are required for [the 7th vignette](https://alstat.github.io/ALUES/articles/a07_visual_maps.html). I was unable to install ggmap due to an outdated R installation, but i was able to knit the vignette file after commenting out both `library(ggmap)` and `library(raster)`. Are these packages indeed necessary, or can they be dropped from the vignette and from the package `Suggests` list? (I apologize if the answer is obvious; i don't use these packages myself.)

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/4228)._
