### 2022-03-15 alstat/ALUES#20 @corybrunson
illustrate non-error output elements in overall_suit documentation
This is a minor point, but possibly due to an oversight so i wanted to flag it. Here is the first chunk of example code for `overall_suit()`:

```r
library(ALUES)
out <- suit("ricebr", terrain=MarinduqueLT, water=MarinduqueWater, temp=MarinduqueTemp, sow_month=1)
out[["terrain"]]
```

Because the crop requirements don't match the terrain data, the third line contains only an error string. Should a fourth line, e.g. `out[["water"]]`, be added to show the user how the output looks when the crop and environment data are compatible?

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/4228)._
