### 2022-03-15 alstat/ALUES#21 @corybrunson
cross-reference documented functions
The package contains a large number of data objects, which makes the index a slog to browse. Because there are only three documented functions, all of which are at least indirectly related, i suggest that each function include a link to both others in the "Description" or "See also" section. This would provide easy navigation among them, which is not possible through the index.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/4228)._
