### 2022-03-15 alstat/ALUES#22 @corybrunson
misspecified test(s)?
While running tests in "test_suitability.R", i encountered a failure on line 50:

```r
test_that("Case A: Triangular", expect_equal(suit$`Suitability Class`["CFragm"][6,], right_tri()[["class"]]))
```

The reason is that one side is a character string while the other side is a factor value.

This might be due to a difference in versions of the testthat package, but i wanted to flag it to check that it's intended and, if not, ask if other similar tests may need to be revised.

(Also, i had to export `suitability()` in order to manually run them, which prompts me to ask: Why is this function documented but not exported? I apologize if i missed an answer somewhere.)

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/4228)._
