+++ README.md
@@ -29,16 +29,19 @@ install_github("alstat/ALUES")
 ```
 We want to hear some feedbacks, so if you have any suggestion or issues regarding this package, please do submit it [here](https://github.com/alstat/ALUES/issues/). As for those who would want to contribute please read the [CONTRIBUTING.md](https://github.com/alstat/ALUES/blob/master/CONTRIBUTING.md) file.
 
-## Authors
+## Authors (with Contribution)
 1. [Al-Ahmadgaid B. Asaad](https://github.com/alstat/) (Maintainer)
     * email: alahmadgaid@gmail.com
     * website: https://al-asaad.github.io/
+    * role: main developer of the package
 <br><br>
 2. [Arnold R. Salvacion](https://github.com/arsalvacion/)
     * email: arsalvacion@gmail.com
     * blog: http://r-nold.blogspot.com/
+    * role: supervisor and domain experts on agriculture, and provider of all the necessary articles 
 <br><br>
 3. Bui Tan Yen
+    * role: developer of the LUSET, which was the basis of ALUES
     
 
 
