+++ tests/testthat/test_suitability_unbias.R
@@ -2,18 +2,18 @@ library(testthat)
 library(ALUES)
 
 # NA PRESENCE
-suit <- suitability(LaoCaiLT, ALFALFASoil, interval="unbias")
-test_that("Parameter name present:", expect_equal(suit$`Factors Evaluated`[1], "SoilTe"))
+suit_ <- suitability(LaoCaiLT, ALFALFASoil, interval="unbias")
+test_that("Parameter name present:", expect_equal(suit_$`Factors Evaluated`[1], "SoilTe"))
 test_that("Parameter scores NA all:", 
-          expect_equal(sum(is.na(suit$`Suitability Score`["SoilTe"])), 
-          nrow(suit$`Suitability Score`["SoilTe"])))
+          expect_equal(sum(is.na(suit_$`Suitability Score`["SoilTe"])), 
+          nrow(suit_$`Suitability Score`["SoilTe"])))
 test_that("Parameter classes NA all:", 
-          expect_equal(sum(is.na(suit$`Suitability Class`["SoilTe"])), 
-                       nrow(suit$`Suitability Class`["SoilTe"])))
+          expect_equal(sum(is.na(suit_$`Suitability Class`["SoilTe"])), 
+                       nrow(suit_$`Suitability Class`["SoilTe"])))
 test_that("Parameter Minimum NA:", 
-          expect_true(is.na(suit$`Factors' Minimum Values`["SoilTe"])))
+          expect_true(is.na(suit_$`Factors' Minimum Values`["SoilTe"])))
 test_that("Parameter Maximum NA:", 
-          expect_true(is.na(suit$`Factors' Maximum Values`["SoilTe"])))
+          expect_true(is.na(suit_$`Factors' Maximum Values`["SoilTe"])))
 
 # WARNINGS
 test_that("Expecting Warning", expect_warning(suitability(LaoCaiLT, ALFALFASoil)))
@@ -24,7 +24,7 @@ test_that("Expecting Warning", expect_warning(suitability(LaoCaiLT, ALFALFASoil)
 # Right Face Triangular MF
 
 LaoCaiLT2 <- LaoCaiLT[5:6,]
-suit <- suitability(LaoCaiLT2, SOYASoil, interval="unbias")
+suit_ <- suitability(LaoCaiLT2, SOYASoil, interval="unbias")
 right_tri <- function (r) {
   x <- LaoCaiLT2[r,"CFragm"]; Min <- 0
   reqScore <- as.numeric(SOYASoil[1,2:7])
@@ -46,15 +46,15 @@ right_tri <- function (r) {
   }
   return(list("score" = score, "class" = class_))
 }
-test_that("Case A: Triangular", expect_equal(suit$`Suitability Score`["CFragm"][1,], right_tri(1)[["score"]]))
-test_that("Case A: Triangular", expect_equal(suit$`Suitability Class`["CFragm"][1,], right_tri(1)[["class"]]))
-test_that("Case A: Triangular", expect_equal(suit$`Suitability Score`["CFragm"][2,], right_tri(2)[["score"]]))
-test_that("Case A: Triangular", expect_equal(suit$`Suitability Class`["CFragm"][2,], right_tri(2)[["class"]]))
+test_that("Case A: Triangular", expect_equal(suit_$`Suitability Score`["CFragm"][1,], right_tri(1)[["score"]]))
+test_that("Case A: Triangular", expect_equal(suit_$`Suitability Class`["CFragm"][1,], right_tri(1)[["class"]]))
[…]
