+++ tests/testthat/test_suitability_min.R
@@ -7,7 +7,7 @@ library(ALUES)
 # Right Face Triangular MF
 
 LaoCaiLT2 <- LaoCaiLT[5:6,]
-suit <- suitability(LaoCaiLT2, SOYASoil, interval="unbias", min = "average", max = 20)
+suit_ <- suitability(LaoCaiLT2, SOYASoil, interval="unbias", min = "average", max = 20)
 right_tri <- function (r) {
   x <- LaoCaiLT2[r,"CFragm"]
   reqScore <- as.numeric(SOYASoil[1,2:7])
@@ -17,14 +17,14 @@ right_tri <- function (r) {
   return(c(Min, 20))
 }
 
-test_that("Case A: Triangular", expect_equal(as.numeric(suit$`Factors' Minimum Values`["CFragm"]), right_tri(1)[1]))
-test_that("Case A: Triangular", expect_equal(as.numeric(suit$`Factors' Maximum Values`["CFragm"]), right_tri(1)[2]))
+test_that("Case A: Triangular", expect_equal(as.numeric(suit_$`Factors' Minimum Values`["CFragm"]), right_tri(1)[1]))
+test_that("Case A: Triangular", expect_equal(as.numeric(suit_$`Factors' Maximum Values`["CFragm"]), right_tri(1)[2]))
 
-suit <- suitability(LaoCaiLT2, SOYASoil, interval="unbias", min = -1)
-test_that("Case A: Triangular", expect_equal(as.numeric(suit$`Factors' Minimum Values`["CFragm"]), -1))
+suit_ <- suitability(LaoCaiLT2, SOYASoil, interval="unbias", min = -1)
+test_that("Case A: Triangular", expect_equal(as.numeric(suit_$`Factors' Minimum Values`["CFragm"]), -1))
 # 
-# suit <- suitability(LaoCaiLT2, SOYASoil, interval="unbias", min = rep(-1, 10))
-# test_that("Case A: Triangular", expect_equal(as.numeric(suit$`Factors' Minimum Values`), rep(-1, 7)))
+# suit_ <- suitability(LaoCaiLT2, SOYASoil, interval="unbias", min = rep(-1, 10))
+# test_that("Case A: Triangular", expect_equal(as.numeric(suit_$`Factors' Minimum Values`), rep(-1, 7)))
 # test_that("Case A: Triangular", expect_error(suitability(LaoCaiLT2, SOYASoil, interval="unbias", min = rep(-1, 5))))
 
 # ------------------------------
@@ -34,7 +34,7 @@ test_that("Case A: Triangular", expect_equal(as.numeric(suit$`Factors' Minimum V
 
 LaoCaiLT3 <- LaoCaiLT[5:6,]
 LaoCaiLT3[1,2] <- 80
-suit <- suitability(LaoCaiLT3, SOYASoil[8:nrow(SOYASoil),], interval = "unbias", min = "average", max = 20)
+suit_ <- suitability(LaoCaiLT3, SOYASoil[8:nrow(SOYASoil),], interval = "unbias", min = "average", max = 20)
 left_tri <- function(r) {
   x <- LaoCaiLT3[r,"OC"]
   reqScore <- as.numeric(SOYASoil[8,2:7])
@@ -44,21 +44,21 @@ left_tri <- function(r) {
   return(c(Min, Max))
 }
 
-test_that("Case A: Triangular", expect_equal(as.numeric(suit$`Factors' Minimum Values`["OC"]), left_tri(1)[1]))
-test_that("Case A: Triangular", expect_equal(as.numeric(suit$`Factors' Maximum Values`["OC"]), left_tri(1)[2]))
[…]
