+++ DESCRIPTION
@@ -36,7 +36,7 @@ Depends:
     Rcpp (>= 0.10.6)
 Suggests:
     testthat, markdown, knitr, microbenchmark, ggmap, raster, reshape2
-RoxygenNote: 7.1.1
+RoxygenNote: 7.1.2
 VignetteBuilder: knitr
 URL: https://github.com/alstat/ALUES/
 BugReports: https://github.com/alstat/ALUES/issues/

[…]
