### 2022-04-28 alstat/ALUES#22 @alstat
Hi @corybrunson, I not getting an error. Although, I made some fix to the tests, I changed the `suit` variable to `suit_` to avoid conflicting with the exported function `suit`. In case, this might be the reason for errors.

As for the `suitability`, we initially used it as our exported function for computing the suitability score, but we decided to shorten it to `suit` to make it consistent with the `overall_suit` function. So, basically `suit` is a wrapper for `suitability`.
