### 2022-04-28 3676774
re-organized figures and added link to other functinos
