### 2022-04-28 alstat/ALUES#21 @alstat
Added this to See also section. There are only two functions exported, `suit` and `overall_suit` both now have link to each other. Thank you for this suggestion.
