### 2022-04-28 5218b75
add example for output with compatible crops
