### 2022-04-28 alstat/ALUES#19 @alstat
Hi @corybrunson, both are required. The reason why the knitting works without it, follows from the fact that the figures generated for the maps are actually saved images. We don't really run the code to generate the images because it would take time, so we decided to save the output and simply call it into the vignette. However, if you need to run the code in the console, by copy-pasting, then you really have to load both `ggmap` and `raster`.
