### 2022-04-28 alstat/ALUES#22 @corybrunson
Hm. I just tried to reproduce the error without updating the repo, so this might have been my fault. Thank you for the explanation!

Regarding `suitability()`, even as a wrapper, it would be good to just include it in the documentation for `suit()`, if it's going to be exported. But it wouldn't require anything more that a sentence saying that it's a wrapper. Is there a specific reason not to?
