### 2022-04-29 alstat/ALUES#22 @alstat
Thank you @corybrunson, the specific reason as to why `suitability()` function is not documented in `suit()` is the fact that it is not an exported function.
