+++ README.md
@@ -14,6 +14,9 @@ uses fuzzy logic approach to evaluate land suitability of a particular area base
 temperature, topography, and soil properties. The membership functions used for fuzzy modeling are the
 following: _Triangular_, _Trapezoidal_ and _Gaussian_. The methods for computing the overall suitability of a particular area are also included, and these are the _Minimum_, _Maximum_ and _Average_. Finally, ALUES is a highly optimized library with core algorithms written in C++.
 
+## Statement of Need
+Several computer systems have been developed for agricultural land suitability assessments. Examples of these are ALES ([Johnson & Cramb, 1991](https://doi.org/10.1111/j.1475-2743.1991.tb00881.x)), LEIGIS ([Kalogirou, 2002](https://doi.org/10.1016/S0198-9715(01)00031-X)) and ALES ([Elsheikh et al., 2013](https://doi.org/10.1016/j.compag.2013.02.003)). However, most of these systems are either proprietary or lacking features for land suitability assessments. It is therefore the goal of this software, ALUES, to address some the limitations of the aforementioned software.
+
 ## Installation
 The package is available on CRAN, and can be installed as follows:
 
