### 2022-04-29 alstat/ALUES#17 @alstat
Added Statement of Need to README. Commit [02344b2](https://github.com/alstat/ALUES/commit/02344b25f58b7ff6d2782a2b2383903f3695ab95)
