### 2022-04-29 96b6134
Change efficient enough to simply benchmark
