### 2022-04-29 alstat/ALUES#16 @alstat
Hi @corybrunson, thank you for your suggestion. I removed the claim of "efficiency" and simply put a benchmark title. See commit [here](https://github.com/alstat/ALUES/commit/96b6134c8122e02baa3d5be2cf818297a603e59c). Should I re-upload the paper to update it?
