### 2022-04-29 87bc1a4
added more docs to suit and overall_suit
