+++ man/suit.Rd
@@ -87,6 +87,92 @@ This function calculates the suitability scores and class of the land units.
 }
 \examples{
 library(ALUES)
+banana_suit <- suit("banana", terrain=MarinduqueLT)
+names(banana_suit)
+# The warning above simply tells the user that one of the factor, CECc, in the target crop requirement, has parameter intervals for all suitability classes equal to 16, and the package used this value as the maximum constant for computing the suitability scores. For more, please refer to the **Article 2: Methodology used in ALUES** of the documentation.
+# The `suit` function returns a list of output of target characteristics, in this case `"terrain"` and `"soil"`. To access the output, simply run the following:
+lapply(banana_suit[["terrain"]], function(x) head(x)) # lapply is used to display the head of each items in the list
+lapply(banana_suit[["soil"]], function(x) head(x)
+# Each of these are lists, with the following names:
+names(banana_suit[["soil"]])
+# So that, to access the factors evaluated, simply run the following:
+banana_suit[["soil"]][["Factors Evaluated"]]
+
+## Targetting Crop
+# There are 56 crops available in ALUES, and what we've illustrated above is for banana only. Other crops are listed below:
+d <- utils::data(package = "ALUES")
+alues_data <- d$results[, "Item"]
+crop_data <- regmatches(alues_data, gregexpr(paste0("^[A-Z]{2,}", collapse = "|"), alues_data))
+crop_data <- unique(unlist(lapply(crop_data, function(x) substr(x, 1, nchar(x)-1))))
+crop_data
+
+# These are the names for the input string for the `suit` function. For example, to target sweet potato the input string is not `"sweet potato"` but rather `potatosw`. That is,
+potato_suit1 <- suit("sweet potato", terrain=MarinduqueLT)
+potato_suit2 <- suit("potatosw", terrain=MarinduqueLT)
+
+## Targetting Crop Factors
+# The idea of evaluating a land unit is to match the quality of the land against the standard value of the target factor. Therefore, if the crop does not include the factor you are targeting, then there won't be any matching to be done. For example, the land units evaluated above are those in Marindque, which has the following soil and terrain characteristics:
+head(MarinduqueLT)
+# The crop that we are trying to target is banana. The `suit` function simply require the user to input a string name for the target crop, and the function will look for the corresponding crop datasets. For example, for banana these are the crop requirements datasets for the four characteristics:
+BANANATerrain
+BANANASoil
+BANANAWater
+BANANATemp
+# These datasets are used by the `suit` function depending on the targetted characteristics of the input land units (specified by the user) on the said function. So for `banana_suit` object above, the target crop datasets were `BANANATerrain` and `BANANASoil` since the input land unit specified is `terrain=MarinduqueLT`. Further, the input land unit only targetted the soil factors and not the terrain factors, since none of the factors in `MarinduqueLT` matched with the factors in `BANANATerrain`. That is why, accessing the output for the terrain characteristics for the `banana_suit` object will return the following:
+banana_suit[["terrain"]]
+
+## Targetting Multiple Characteristics
+# The example above only targetted the terrain and soil characteristics, but the `suit` function allows user to also target water and temp simultaneously. For examples, we can evaluate the land units of Lao Cai, Vietnam for all three characteristics as follows for irrigated rice (`riceiw``):
+riceiw_multi <- suit("riceiw", terrain=LaoCaiLT, water=LaoCaiWater, temp=LaoCaiTemp, sow_month=10)
+names(riceiw_multi)
[…]

[…]
