### 2022-04-29 alstat/ALUES#18 @alstat
I have updated the vignette to printout the `banana_suit[["terrain"]]` and added a few explanations to help the readers. Further, I also added the docs from the vignette on "Understanding suit" and "Understanding overal_suit" to their respective R docs for `suit()` and `overall_suit()`. Check the commit [here](https://github.com/alstat/ALUES/commit/87bc1a432fc4ecfdfa4b863c01b5801407445224). Thanks.
