+++ R/suit.R
@@ -68,92 +68,186 @@
 #' library(ALUES)
 #' banana_suit <- suit("banana", terrain=MarinduqueLT)
 #' names(banana_suit)
-#' # The warning above simply tells the user that one of the factor, CECc, in the target crop requirement, has parameter intervals for all suitability classes equal to 16, and the package used this value as the maximum constant for computing the suitability scores. For more, please refer to the **Article 2: Methodology used in ALUES** of the documentation.
-#' # The `suit` function returns a list of output of target characteristics, in this case `"terrain"` and `"soil"`. To access the output, simply run the following:
-#' lapply(banana_suit[["terrain"]], function(x) head(x)) # lapply is used to display the head of each items in the list
-#' lapply(banana_suit[["soil"]], function(x) head(x)
+#' 
+#' # The warning above simply tells the user that one of the factor, 
+#' # CECc, in the target crop requirement, has parameter intervals for 
+#' # all suitability classes equal to 16, and the package used this value 
+#' # as the maximum constant for computing the suitability scores. For more,
+#' # please refer to the **Article 2: Methodology used in ALUES** of the documentation.
+#' # The `suit` function returns a list of output of target 
+#' # characteristics, in this case `"terrain"` and `"soil"`. To access 
+#' # the output, simply run the following:
+#'
+#' # lapply is used to display the head of each items in the list
+#' lapply(banana_suit[["terrain"]], function(x) head(x)) 
+#' lapply(banana_suit[["soil"]], function(x) head(x))
+#' 
 #' # Each of these are lists, with the following names:
 #' names(banana_suit[["soil"]])
+#' 
 #' # So that, to access the factors evaluated, simply run the following:
 #' banana_suit[["soil"]][["Factors Evaluated"]]
 #' 
 #' ## Targetting Crop
-#' # There are 56 crops available in ALUES, and what we've illustrated above is for banana only. Other crops are listed below:
+#' # There are 56 crops available in ALUES, and what we've illustrated 
+#' # above is for banana only. Other crops are listed below:
 #' d <- utils::data(package = "ALUES")
 #' alues_data <- d$results[, "Item"]
-#' crop_data <- regmatches(alues_data, gregexpr(paste0("^[A-Z]{2,}", collapse = "|"), alues_data))
-#' crop_data <- unique(unlist(lapply(crop_data, function(x) substr(x, 1, nchar(x)-1))))
+#' crop_data <- regmatches(alues_data, gregexpr(paste0("^[A-Z]{2,}", 
+#' collapse = "|"), alues_data))
+#' crop_data <- unique(unlist(lapply(crop_data, 
+#' function(x) substr(x, 1, nchar(x)-1))))
 #' crop_data
 #' 
-#' # These are the names for the input string for the `suit` function. For example, to target sweet potato the input string is not `"sweet potato"` but rather `potatosw`. That is,
-#' potato_suit1 <- suit("sweet potato", terrain=MarinduqueLT)
+#' # These are the names for the input string for the 
+#' # `suit` function. For example, to target sweet potato the 
+#' # input string is not `"sweet potato"` but rather `potatosw`. That is,
+#' # potato_suit1 <- suit("sweet potato", terrain=MarinduqueLT)
 #' potato_suit2 <- suit("potatosw", terrain=MarinduqueLT)
 #' 
 #' ## Targetting Crop Factors
-#' # The idea of evaluating a land unit is to match the quality of the land against the standard value of the target factor. Therefore, if the crop does not include the factor you are targeting, then there won't be any matching to be done. For example, the land units evaluated above are those in Marindque, which has the following soil and terrain characteristics:
+#' # The idea of evaluating a land unit is to match the 
[…]
