### 2022-04-30 #4228 @alstat
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1016/j.envsoft.2003.10.006 is OK
> - 10.1016/j.compag.2013.02.003 is OK
> - 10.1080/10618600.1996.10474713 is OK
> - 10.1111/j.1475-2743.1991.tb00881.x is OK
>
> MISSING DOIs
>
> - None
>
> INVALID DOIs
>
> - 10.1016/S0198-97150100031-X is INVALID
> - 10.1016/S0019-99586590241-X is INVALID
> ```

The invalid DOIs here are due to missing parentheses. So,
```
10.1016/S0198-97150100031-X # invalid
10.1016/S0198-9715(01)00031-X # correct
```
and,
```
10.1016/S0019-99586590241-X # invalid
10.1016/S0019-9958(65)90241-X # correct
```
The correct DOIs are the ones in the paper, not sure why the parentheses are omitted when checkin. For reference, here are the complete URLs for the two DOIs above:
```
https://doi.org/10.1016/S0198-9715(01)00031-X
https://doi.org/10.1016/S0019-9958(65)90241-X
```
Not sure how to resolve this.
