### 2022-05-02 #4228 @fabian-s
@kuesterc9
you left the checkmarks for "automated tests" and "community guidelines" out -- do you have remaining points of criticism or have your concerns been addressed in the meantime?
