### 2022-05-02 #4228 @fabian-s
@alstat

please take a look at the issue I opened re. best practices for scientific software in R and take a look at the PR suggesting some minor edits to the paper.

Addressing https://github.com/alstat/ALUES/issues/24 is not necessary for JOSS publication. IMO would still be a good idea that would improve your package.
