### 2022-05-03 alstat/ALUES#24 @alstat
Noted on these @fabian-s, thank you. These are indeed useful suggestions and will consider these on next release.
