### 2022-05-09 #4228 @alstat
Hi @fabian-s, below is the link of the archived software in Zenodo, following the correct metadata:

https://zenodo.org/record/6529873#.YniNtRNBx-o

Here is the DOI of the archived version: 10.5281/zenodo.6529873
